Use int instead of size_t
This is to make CI not fail (I wonder why it compiles fine on my machine
though...)

Also, why does length() return an int anyways? Wouldn't size_t make more
sense?

+++ core/string/ustring.cpp
@@ -2099,7 +2099,7 @@ int64_t String::hex_to_int(bool p_with_prefix) const {
 		return 0;
 	}
 
-	size_t i = 0;
+	int i = 0;
 
 	int64_t sign = this->ptr()[i] == '-' ? -1 : 1;
 
@@ -2141,7 +2141,7 @@ int64_t String::bin_to_int(bool p_with_prefix) const {
 		return 0;
 	}
 
-	size_t i = 0;
+	int i = 0;
 
 	int64_t sign = this->ptr()[i] == '-' ? -1 : 1;
